Extract session configuration to its own module
